Fix use of `ok` var
Prior to this change, we were checking the `ok` var to see if the
BUILDPULSE_SECRET_ACCESS_KEY env var is present, but we were
accidentally reusing the `ok` var that was populated when checking to
see whether the BUILDPULSE_ACCESS_KEY_ID env var is present.

Because the value is "" when the env var is not present, and because
we were already checking for a value of "", the old code didn't result
in a bug, but it was nevertheless some less-than-ideal code, so this
commit fixes that up.