Gracefully handle nonexistent repository
This functionality should be considered experimental for now, until we
can exercise it in more real world environments. With that in mind,
for any environments that are using the existing production
test-reporter (v0.6.0), we want this new functionality to offer
graceful degradation: if we're unable to find a repository, or if we run
into problems fetching the commit metadata, try to ensure that the
failure doesn't interfere with the other existing functionality.

Once this new functionality is exercised in more real world
environments, if all is well, then we can revert this commit such that
a failure to fetch the git metadata is considered a fatal failure.