Add support for specifying tree SHA via arg
In some scenarios, a local git repository does not exist at the time
when the CI build is reporting test results to BuildPulse (e.g., builds
that use a tarball of the repository instead of perfoming a `git clone`,
or builds that delete the `.git` directory early in the build process to
save space). In these scenarios, since no git repository is present, we
can't use git to look up the tree SHA for the commit.

To support these scenarios, this change introduces an optional `--tree`
arg for specifying the tree SHA when invoking the CLI.