Teach toTar to exclude directories that lack relevant files
Prior to this change, toTar created a directory for *every* directory in
the source dir, regardless of whether that directory contained relevant
files (i.e., buildpulse.yml, buildpulse.log, or *.xml).

With this change in place, toTar only creates directories when those
directories contain one or more of the relevant files.