⬆️ mholt/archiver@074da64920d3
Bump mholt/archiver to pick up the security fix in
https://github.com/mholt/archiver/pull/272/commits/41df2e662af197af79e9,
which resolves GHSA-25xm-hr59-7c27.

Note that there isn't yet a published release of mholt/archiver
including that security fix, so we're pinning mholt/archiver to the most
recent commit that has a passing CI build (074da64920d3). The diff
between the previous release and that commit is available at:

https://github.com/mholt/archiver/compare/v3.5.0...074da64920d3